# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Curriculum Activity & Attendance App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students with optional filters
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: section
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /students/{id}/attendance-summary:
    get:
      operationId: getStudentAttendanceSummary
      tags: [students]
      summary: Get attendance summary for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"

  # FACULTY
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [curriculum]
      summary: List subjects/curriculum
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: semester
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [curriculum]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [curriculum]
      summary: Get subject by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Not found
    patch:
      operationId: updateSubject
      tags: [curriculum]
      summary: Update a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [curriculum]
      summary: Delete a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ACTIVITY RECOMMENDATIONS
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create a notification (admin/faculty)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, registerNumber, name, departmentId, year, section]
      properties:
        id:
          type: integer
        registerNumber:
          type: string
        name:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: ["string", "null"]
        year:
          type: integer
        section:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        attendancePercentage:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    StudentInput:
      type: object
      required: [registerNumber, name, departmentId, year, section]
      properties:
        registerNumber:
          type: string
        name:
          type: string
        departmentId:
          type: integer
        year:
          type: integer
        section:
          type: string
        email:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
    StudentUpdate:
      type: object
      properties:
        registerNumber:
          type: string
        name:
          type: string
        departmentId:
          type: integer
        year:
          type: integer
        section:
          type: string
        email:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
    StudentListResponse:
      type: object
      required: [students, total, page, limit]
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    AttendanceSummary:
      type: object
      required: [studentId, totalClasses, present, absent, late, percentage]
      properties:
        studentId:
          type: integer
        totalClasses:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        percentage:
          type: number
        bySubject:
          type: array
          items:
            $ref: "#/components/schemas/SubjectAttendanceSummary"
    SubjectAttendanceSummary:
      type: object
      required: [subjectId, subjectName, total, present, percentage]
      properties:
        subjectId:
          type: integer
        subjectName:
          type: string
        total:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        percentage:
          type: number
...
    Subject:
      type: object
      required: [id, code, name, department, semester, credits]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        department:
          type: string
        semester:
          type: integer
        credits:
          type: integer
        lectureHours:
          type: ["integer", "null"]
        practicalHours:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SubjectInput:
      type: object
      required: [code, name, department, semester, credits]
      properties:
        code:
          type: string
        name:
          type: string
        department:
          type: string
        semester:
          type: integer
        credits:
          type: integer
        lectureHours:
          type: integer
        practicalHours:
          type: integer
        category:
          type: string
    SubjectUpdate:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        department:
          type: string
        semester:
          type: integer
        credits:
          type: integer
        lectureHours:
          type: integer
        practicalHours:
          type: integer
        category:
          type: string
...
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [low_attendance, new_activity, timetable_change, assignment, general]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        targetRole:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    NotificationInput:
      type: object
      required: [type, title, message]
      properties:
        type:
          type: string
          enum: [low_attendance, new_activity, timetable_change, assignment, general]
        title:
          type: string
        message:
          type: string
        targetRole:
          type: string
...