# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Curriculum Activity & Attendance App API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  code?: string;
  name?: string;
  department?: string;
  semester?: number;
  credits?: number;
  lectureHours?: number;
  practicalHours?: number;
  category?: string;
}